requirements.txt
flask
flask-cors
pypdf
beautifulsoup4
groq
requests
python-dotenv
huggingface_hub
